split into TypeLibs/Abstract/API/Debug files, moved into ComManagement, and updated using statements
